Bug fixes, track time and duration, and end of stream
The main player now takes into account the remaining buffers in the
output buffer before emitting an ‘end_of_stream’ event, and the timer
in the main interface keeps going long enough to reflect the real
duration of the track. Also, the decoder now reflects the real duration
of the track if it is truncated, instead of the full track. Introduced
a ‘track_length’ which holds the length of the full track for the
players, and playback_manager.